ImageController now renders on an independent timer, no call from MatrixController needed (in fact, nothing should ever call ImageController.render() except ImageController.show()). The numbers for ImageController's update timer and target queue depth are made up and could probably be improved.
MatrixController is now MUCH simpler.

Never auto-show when changing a setting (was unnecessary and actually undesirable).

Added "nodisplay" flag for debugging with console output.

Propgate settings changes into img_ctrl object.

Define an initial generic ImageController in hawks.__init__().

Correctly detect a lack of preset.

Keep track of the current imagecontroller object so that we can tell
it to stop rendering (and therefore exit and die) when we make a new one.

The animations that used to happen in MatrixController are still commented
out in ImageController and have not been implemented yet. Also, the
subclasses of ImageController are still rendering their full sets of
static frames at show() time.