Fix a bug where the first frame of a dynamically rendered ImageController child would get eaten.
Add a flag and a behavior for only repeating animated GIFs whole numbers
of times.

Remove time as an explicit factor in SlideShow rendering logic. Frames
are enqueued so as to fill up the necessary time, and the time of day
when the functions are running is not super meaningful.